refactor: change component directory name to PascalCase
