+++ apps/web/src/modules/job/components/card/components/Footer/styled.ts
