+++ apps/web/src/modules/job/components/card/components/Footer/type.ts
