+++ apps/web/src/modules/job/components/card/components/Header/styled.ts
