+++ apps/web/src/modules/job/components/card/components/ShootingDetail/index.tsx
