+++ apps/web/src/modules/job/components/card/components/ShootingDetail/styled.ts
